Update filtering + usage of URL params
Update the filtering on the template demo store to more directly use the
input provided by the storefront API in the URL params. This also
allows us to more directly use the URL params in requests to the API.

It also makes the filtering more robust, as it will now work with
filters based on metafields and metaobjects, not just tags, booleans,
price, and variantOptions.

Also removed the filter values > 1 constraint on showing filters.
I feel that this is a better UX, and this constraint was causing
the price filter to be hidden.