Ensure proper exit code is being used
This previously had a comment of "This code should not be reached" but
in testing it appeared to always be reached. This had the unfortunate
side effect of always returning with an exit code of 0 even if a command
manually called `process.exit(1)`.

Instead of removing the exit from here we can get the result of the exit
code from `spawnSync` and pass it along.