Add scripts to make it easier to see true diffs in recipes (as a reviewer)
Reviewing recipe changes isn't straightforward because the patch files get renamed on top of
being modified, so reviewers can't see a clean diff. The rename-patches-for-review script will
rename all of the patch files to their "old" names, so that reviewers can see a clean diff of
the actual changes.

Also in the `.prompt.md` files, there is a lot of noise that is just a SHA changing from the old
to the new SHA. The review-sha-cleanup script will find and replace all occurrences of the new sha
with the old sha so that all noise is removed and reviewers will only see legitimate code changes
when viewing the `.prompt.md` diffs.

`npm run review-recipe-diffs` runs both scripts. These scripts are NOT intended to be used during
development, only during the review phase. The resulting code changes should NOT be committed and pushed.