Update tsconfig so that `cmd+click`ing takes you to source code, not d.ts file
Previously, if I was in a file like `templates/skeleton/app/root.tsx` and I cmd+clicked
on the function 'getShopifyAnalytics` (to go to its definition), I was taken to a d.ts file and not the actual source code in the Hydrogen package.

Now, I'm taken to the actual source code (ie: `packages/hydrogen/src/analytics-manager/AnalyticsProvider.tsx in this case).